+++ src/devices/Uln2003/README.md
@@ -7,7 +7,7 @@ A stepper motor is an electromechanical device which converts electrical pulses 
 
 The 28BYJ-48 is a small stepper motor suitable for a large range of applications. More information [here](https://components101.com/motors/28byj-48-stepper-motor)
 
-![](SM28BYJ48.png)
+![](Uln2003.png)
 
 **[Stepper Motor 28BYJ-48]**: http://www.geeetech.com/Documents/Stepper%20motor%20datasheet.pdf
