+++ src/devices/Uln2003/Uln2003.csproj
@@ -8,7 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="System.Device.Gpio" Version="0.1.0-prerelease*" />
-    <Compile Include="StepperMotor.cs" />
+    <Compile Include="Uln2003.cs" />
     <Compile Include="StepperMode.cs" />
   </ItemGroup>
